[added] set lightcolor and fixed the display of the lights

+++ app/src/main/java/app/com/huelightsalarm/models/data/Light.java
@@ -74,7 +74,7 @@ public class Light {
     }
 
     public float map(int initialLowerBounds, int initialUpperBounds, int lowerBounds, int upperBounds, int value) {
-        return ((upperBounds-lowerBounds)/(initialUpperBounds - (float) initialLowerBounds))*value;
+        return ((float) (upperBounds-lowerBounds)/ (float) (initialUpperBounds - initialLowerBounds)) * (float) value;
     }
 
     /**
